Move the tests back to their original locations :'(
