+++ elements/bulbs-poll/bulbs-poll.test.js
@@ -1,6 +1,6 @@
 import fetchMock from 'fetch-mock';
 import testElement from 'bulbs-elements/test/element';
-import '../bulbs-poll';
+import './bulbs-poll';
 
 testElement('<bulbs-poll> <BulbsPoll>', function () {
   let setSrcSpy;
