+++ elements/bulbs-poll/components/loading.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { assertJSXEqual } from 'bulbs-elements/test/assertions';
-import Loading from '../../components/loading';
+import Loading from './loading';
 
 describe('<bulbs-poll> <Loading>', function () {
   context('default', function () {
