+++ elements/bulbs-poll/components/selection-marker.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { assertJSXEqual } from 'bulbs-elements/test/assertions';
-import SelectionMarker from '../../components/selection-marker';
+import SelectionMarker from './selection-marker';
 
 describe('<bulbs-poll> <SelectionMarker>', function () {
   context('is selected', function () {
