+++ elements/bulbs-poll/fields/now.test.js
@@ -1,5 +1,5 @@
 import { assert } from 'chai';
-import NowField from '../../fields/now';
+import NowField from './now';
 
 describe('<bulbs-poll> NowField', function () {
