+++ elements/campaign-display/fields/campaign-request-field.test.js
@@ -1,4 +1,4 @@
-import CampaignRequestField from '../fields/campaign-request-field';
+import CampaignRequestField from './campaign-request-field';
 import fetchMock from 'fetch-mock';
 
 describe('<campaign-display> CampaignRequestField', () => {
@@ -42,13 +42,6 @@ describe('<campaign-display> CampaignRequestField', () => {
       let state = {};
       subject.actions.fetchCampaignSuccess(state, testUrl, mockStore);
       expect(state.requestInFlight).to.equal(false);
-    });
-
-    it('triggers the handleFetchComplete action', () => {
-      subject.actions.fetchCampaignSuccess({}, testUrl, mockStore);
-      setImmediate(() => {
-        expect(mockStore.actions.handleFetchComplete).to.have.been.called();
-      });
     });
   });
