+++ lib/bulbs-elements/components/cropped-image.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { createRenderer } from 'react-addons-test-utils';
-import CroppedImage from '../components/cropped-image';
+import CroppedImage from './cropped-image';
 
 describe('<cropped-image>', () => {
   let subject;
