+++ lib/bulbs-elements/util/iterate-object.test.js
@@ -1,4 +1,4 @@
-import iterateObject from '../util/iterate-object';
+import iterateObject from './iterate-object';
 
 describe('iterateObject', () => {
   let subject;
